+++ R/trajectory.R
@@ -33,16 +33,17 @@
 .checkCoords <- function(coords) {
   # Remove NA values at the start or end of the coordinates
   for (startIdx in 1:nrow(coords)) {
-    if (!anyNA(coords[startIdx,])) break
+    cols <- c("x", "y", "time")
+    if (!anyNA(coords[startIdx, cols])) break
   }
   for (endIdx in nrow(coords):1) {
-    if (!anyNA(coords[endIdx,])) break
+    if (!anyNA(coords[endIdx, cols])) break
   }
   coords <- coords[startIdx:endIdx,]
   # Shouldn't be any NAs remaining
-  if (anyNA(coords)) {
-    stop(sprintf("Trajectory contains missing coordinate values, first row with NA is %d",
-                 which(is.na(coords))[1]))
+  if (anyNA(coords[, cols])) {
+    stop(sprintf("Trajectory contains missing coordinate or time values, first row with NA is %d",
+                 which(is.na(coords[, cols]), arr.ind = TRUE)[1, 1] + startIdx - 1))
   }
   coords
 }
@@ -53,7 +54,10 @@
 #'
 #' \code{TrajFromCoords} creates a new trajectory object from a set of
 #' 2-dimensional cartesian coordinates, times and some metadata. The coordinates
-#' are sometimes referred to as "relocations".
+#' are sometimes referred to as "relocations". Rows containing \code{NA}
+#' coordinate or time values at the start or end of the trajectory are
+#' discarded. \code{NA} coordinates or times in the middle of the trajectory
+#' generate an error.
 #'
 #' If \code{timeCol} is specified, \code{track[,timeCol]} is expected to contain
 #' the time (in some numeric units) of each coordinate. Otherwise, times are
@@ -86,8 +90,8 @@
 #'   the following components: \item{x}{X coordinates of trajectory points.}
 #'   \item{y}{Y coordinates of trajectory points.} \item{time}{Time (in
 #'   \code{timeUnits}) for each point. if \code{timeCol} is specified, values
-#'   are \code{track[,timeCol]}, otherwise values are calculated from \code{fps}.}
-#'   \item{displacementTime}{Relative frame/observation times, with
+#'   are \code{track[,timeCol]}, otherwise values are calculated from
+#'   \code{fps}.} \item{displacementTime}{Relative frame/observation times, with
 #'   frame/observation 1 at time \code{0}.} \item{polar}{Coordinates represented
 #'   as complex numbers, to simplify working with segment angles.}
 #'   \item{displacement}{Displacement vectors (represented as complex numbers)